requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
da-vinci==0.4.0
Django==5.0.2
django-admin-thumbnails==0.2.8
django-adminlte-3==0.1.6
django-ckeditor==6.7.0
django-js-asset==3.0.1
django-mptt==0.16.0
django-multiselectfield==0.1.13
django-shortuuidfield==0.1.3
django-thumbnails==0.7.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
requests==2.32.3
rsa==4.9
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
